requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0

# API Server
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0

# Data Processing (scipy 제거 - 패딩 방식 사용)
# scipy>=1.7.0  # 패딩 방식으로 대체
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.62.0

# Utilities
python-multipart>=0.0.5
requests>=2.28.0
pyserial>=3.5
